requirements.txt
numpy==1.21.2
opencv-python==4.5.3.56
dlib==19.22.0
Pillow==8.3.2
PyMuPDF==1.18.19
